m68k-virt: fix ISR register corruption found in final review
The final whole-branch review found a Critical bug in
goldfish_rtc_isr.S: it restored a0 from the stack, then immediately
clobbered it to compute the tail-jump target into int_timerc
(bios/arch/m68k/vectors.S), and since that tail-jump never returns,
a0 was never correctly restored before control passed back to
whatever was interrupted -- corrupting it on every tick, including
every tick before init_system_timer() runs (the no_tick path was
equally broken, unconditionally). No task-scoped review could see
this: it required reading a file (vectors.S) neither owned by nor
touched by the task under review.

Fixed with the frame-synthesis technique bios/arch/m68k/amiga2.S
already establishes as precedent for exactly this problem (a tail
call into a routine ending in its own RTE), simplified since this
board is always 68020+ (no _longframe branch needed).

Empirically tested the reviewer's hypothesis that this corruption
was the actual cause of Task 3's "hz_200 freezes at 2" finding: after
the fix, hz_200 still reaches exactly 2 and stops, unchanged. This
refutes the hypothesis and leaves the original conclusion (a separate,
pre-existing, shared AES/VDI dispatch issue, not part of this branch)
standing on firmer ground.

Also from the same review: goldfish_pic_enable() takes WORD not int;
goldfish_rtc_init() uses the existing VEC_LEVEL6 macro instead of raw
address arithmetic; a clarifying comment in bios/serport.c for the
missing Goldfish TTY init call; ENTRY(_os_entry) added to emutos.ld
for both QEMU virt boards (removes a silent-failure-mode dependency
on link order); and a doc/install.txt section (CLAUDE.md names this
file, not readme.md, as authoritative for "every configuration").

Re-verified: full build succeeds, boots identically (hz_200=2, one
benign FPU-probe trap, VDI/AES console lines), rpi2/virt-arm
portability builds still succeed with ENTRY() added.

+++ bios/machine/virt-m68k/goldfish_pic.c
@@ -29,13 +29,13 @@
 
 void goldfish_pic_init(void)
 {
-    int n;
+    WORD n;
 
     for (n = 0; n < GOLDFISH_PIC_COUNT; n++)
         PIC_IRQ_DISABLE_ALL(n) = 1;   /* value is ignored; any write disables all 32 lines */
 }
 
-void goldfish_pic_enable(int pic, int bit)
+void goldfish_pic_enable(WORD pic, WORD bit)
 {
     PIC_ENABLE(pic) = (1UL << bit);
 }
